requirements.txt
alabaster==0.7.10
Babel==2.4.0
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
#-e git+https://github.com/khanhruby/XmlConvertTool.git@0a66aa158a47564fa11ff7aa99ffd3d7d73890ed#egg=Django
django-debug-toolbar==1.8
django-excel==0.0.6
docutils==0.13.1
et-xmlfile==1.0.1
futures==3.1.1
google-auth==1.0.2
google-cloud-core==0.27.1
google-cloud-translate==1.2.0
googleapis-common-protos==1.5.2
goslate==1.5.1
gunicorn==19.7.1
idna==2.6
imagesize==0.7.1
jdcal==1.3
Jinja2==2.9.6
lxml==3.8.0
MarkupSafe==1.0
mysqlclient==1.3.10
odfpy==1.3.4
olefile==0.44
openpyxl==2.4.8
Pillow==4.1.1
protobuf==3.4.0
pyasn1==0.3.4
pyasn1-modules==0.1.4
pyexcel==0.4.5
pyexcel-io==0.3.4.1
pyexcel-ods==0.3.3
pyexcel-webio==0.0.11
pyexcel-xls==0.3.3
pyexcel-xlsx==0.3.0
Pygments==2.2.0
PyMySQL==0.7.11
pytz==2017.2
requests==2.18.4
rsa==3.4.2
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.2
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.3
texttable==0.9.0
urllib3==1.22
uWSGI==2.0.15
xlrd==1.0.0
xlwt==1.2.0
xmlschema==0.9.9
